requirements.txt
pandas==1.0.3
argparse==1.4.0
requests==2.23.0
python-magic==0.4.18
